# Conflicts:
#	SBS_Ecommerce/SBS_Ecommerce/Content/Theme/Theme3/css/style.css

+++ SBS_Ecommerce/SBS_Ecommerce/Models/OrderDetail.cs
@@ -14,7 +14,7 @@ namespace SBS_Ecommerce.Models
     
     public partial class OrderDetail
     {
-        public double Id { get; set; }
+        public int Id { get; set; }
         public string OrderId { get; set; }
         public int ProId { get; set; }
         public int Quantity { get; set; }
